% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_annual}
\alias{aggregate_annual}
\title{Aggregate into annual means (even precip)}
\usage{
aggregate_annual(x)
}
\arguments{
\item{x}{tibble of GHCNd station data}
}
\value{
monthly summary (by station) tibble
}
\description{
Aggregate into annual means (even precip)
}
