% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_monthly}
\alias{aggregate_monthly}
\title{Aggregate into monthly means (even precip)}
\usage{
aggregate_monthly(x)
}
\arguments{
\item{x}{tibble of GHCNd station data}
}
\value{
monthly summary (by station) tibble
}
\description{
Aggregate into monthly means (even precip)
}
