% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{anomaly}
\alias{anomaly}
\title{Compute anomalies for numeric columns in a grouped table}
\usage{
anomaly(x)
}
\arguments{
\item{x}{tibble}
}
\value{
the input tibble transformed into per group anomalies (anom = value - mean(value))
}
\description{
Compute anomalies for numeric columns in a grouped table
}
