% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{complete_intervals}
\alias{complete_intervals}
\title{Clip a table so that only complete intervals are present (for subsequent aggregation)}
\usage{
complete_intervals(x, by = c("year", "month")[1], min_count = 364)
}
\arguments{
\item{x}{tibble with 'STATION' and 'DATE' variable}

\item{by}{character, one of "year" (default) or "month"}

\item{min_count}{numeric defaults to 364, but adjust to 28 for month}
}
\value{
tibble clipped to include only complete intervals
}
\description{
Clip a table so that only complete intervals are present (for subsequent aggregation)
}
