% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut.R
\name{element_lut}
\alias{element_lut}
\title{Retrieve a look up table describing weather elements}
\usage{
element_lut()
}
\value{
tibble of name, fun (for conversion), mult, units and description
}
\description{
Retrieve a look up table describing weather elements
}
