% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut.R
\name{element_transform}
\alias{element_transform}
\title{Transform variables in a GHCNd station dataset as appropriate}
\usage{
element_transform(x, lut = element_lut())
}
\arguments{
\item{x}{tibble of GHCNd station data}

\item{lut}{tibble, look up table for element transforms}
}
\description{
Transform variables in a GHCNd station dataset as appropriate
}
