% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{fetch_inventory}
\alias{fetch_inventory}
\title{Fetch the inventory listing}
\usage{
fetch_inventory(
  uri = "https://www.ncei.noaa.gov/pub/data/ghcn/daily/ghcnd-inventory.txt",
  add_duration = TRUE,
  stations = NULL
)
}
\arguments{
\item{uri}{char the uri of the inventory}

\item{add_duration}{logical, if TRUE add \code{duration} as a variable}

\item{stations, }{null or the output of \code{fetch_station_list}.  If provided,
then station name is added to the output and the output is filter to include only 
matches in id.  Otherwise, only id is available.}
}
\value{
tibble
}
\description{
VII. FORMAT OF "ghcnd-inventory.txt"
------------------------------
  Variable   Columns   Type
------------------------------
ID           1-11   Character
LATITUDE     13-20   Real
LONGITUDE    22-30   Real
ELEMENT      32-35   Character
FIRSTYEAR    37-40   Integer
LASTYEAR     42-45   Integer
------------------------------
}
