% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{fetch_station}
\alias{fetch_station}
\title{Fetch one or more station(s)}
\usage{
fetch_station(
  id = c("USW00014764", "USW00094623", "USC00172426", "USC00170100"),
  form = c("sf", "raw")[1],
  bind = TRUE,
  transform = TRUE,
  drop_attributes = TRUE,
  baseuri = ghcnd_uri("access"),
  vars = c("STATION", "DATE", "ELEVATION", "NAME", "PRCP", "SNOW", "SNWD", "TMAX",
    "TMIN")
)
}
\arguments{
\item{id}{char, the station identifier(s) or a ghcnd_station_list or ghncd_inventory class}

\item{form}{char, one of 'sf' or 'raw' for raw text}

\item{bind}{logical, if TRUE try to bind the sf/text objects by row.  Be advised, this might
not always work.}

\item{transform}{logical, if TRUE then transform values to real units}

\item{drop_attributes}{logical, if TRUE drop variables that end with the name "_ATTRIBUTES"}

\item{baseuri}{char, the base uri}

\item{vars}{char, if character, the names of variables to keep. If "all"
then keep all.}
}
\value{
list of sf objects or text vectors, one per station OR if \code{bind = TRUE}
  then a single sf object or text vector
}
\description{
Natively returns a list of sf objects, potentially one per station id. Why always
a list?  Because stations may have differing fields.
}
