% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{fetch_station_list}
\alias{fetch_station_list}
\title{Retrieve the GHCNd station list}
\usage{
fetch_station_list(bbox = NULL, baseuri = ghcnd_uri("doc"))
}
\arguments{
\item{bbox}{st_bbox, sfc, sf or NULL, used to subset the returned stations}

\item{baseuri}{char, the baseuri}
}
\value{
tibble or sf of ghcnd_station_list class
}
\description{
------------------------------
Variable   Columns   Type
------------------------------
ID            1-11   Character
LATITUDE     13-20   Real
LONGITUDE    22-30   Real
ELEVATION    32-37   Real
STATE        39-40   Character
NAME         42-71   Character
GSN FLAG     73-75   Character
HCN/CRN FLAG 77-79   Character
WMO ID       81-85   Character
}
