% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access.R
\name{ghcnd_uri}
\alias{ghcnd_uri}
\title{Retrieve the base URI for GHCNd}
\usage{
ghcnd_uri(...)
}
\arguments{
\item{...}{char, path segments to append to the base uri}
}
\value{
char URI
}
\description{
Retrieve the base URI for GHCNd
}
