% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut.R
\name{station_lut}
\alias{station_lut}
\title{Generate a look up table for id, name and geometry}
\usage{
station_lut(x)
}
\arguments{
\item{x}{tibble of GHCNd station data}
}
\value{
table of unique identifiers
}
\description{
Generate a look up table for id, name and geometry
}
