% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut.R
\name{wx_num}
\alias{wx_num}
\title{Conversion function for weather elements of type}
\usage{
wx_num(x, mult = 1)
}
\arguments{
\item{x}{numeric vector}

\item{mult}{numeric multiplier}
}
\value{
modified input vector
}
\description{
Conversion function for weather elements of type
}
