% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut.R
\name{wx_tim}
\alias{wx_tim}
\title{Conversion function for weather elements of type hhmm}
\usage{
wx_tim(x, mult = 1)
}
\arguments{
\item{x}{numeric vector of hhmm}

\item{mult}{ignored}
}
\value{
time in fractional hours
}
\description{
Conversion function for weather elements of type hhmm
}
