% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut.R
\name{wx_txt}
\alias{wx_txt}
\title{Conversion function for weather elements of type text}
\usage{
wx_txt(x, mult = 1)
}
\arguments{
\item{x}{numeric vector of codes}

\item{mult}{ignored}
}
\value{
code as 2-digit character
}
\description{
Conversion function for weather elements of type text
}
